Split Trumpatron Logic Into Separate Libraries
To reduce code clutter, the main trumpatron.py script has been converted to a minimal file, moving most of the logic to the backend modules. The two  modules now in use are bannon.py and trump.py (formerly bot.py).
For the time being, pytest unit tests will be failing. This is planned to be remediated in future commits.